requirements.txt
python-dotenv
azure-cli-core
azure-cli